Use floats instead of ints on multiplication

+++ lib/badmath/badmath.h
@@ -26,6 +26,6 @@
 
 int add(int a, int b);
 int sub(int a, int b);
-int mul(int a, int b);
+float mul(float a, float b);
 
 #endif /* __BADMATH_H */
